+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to LazyProf will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Recipe search in Planning Window - type to filter all profession recipes by name, click to view details
+
 ## [0.5.0] - 2026-02-19
 
 ### Changed